feat: add local storage driver
This commit introduces a new local storage driver to the application.
This allows users to store files and metadata on the local filesystem
instead of relying on cloud-based storage solutions like Azure.

The local storage driver provides the following features:

- File storage: Files are stored in a configurable local directory.
- Metadata storage: File metadata, including content type and user-defined tags, is stored in a separate .metadata directory.
- Directory listing: The driver can list files and directories within the storage path.
- Configuration: The local storage path can be configured in the config.toml file.

This change improves the flexibility and versatility of the application,
making it easier for users to deploy and test it in different environments,
especially self-hosted setup.

Signed-off-by: Denys Fedoryshchenko <denys.f@collabora.com>